Updates get_entities paging.
Changes the expected type from Pager to list. This allows for 'type' to
be combined with a lack of UUID in order to return all objects of a
given type while respecting paging. Updates the several functions which
rely on get_entities to specify that 'fetch_all' should be added to the
kwargs if a UUID is omitted.