Updates request for empty response.
Modifies the 'request' method to handle an empty response when a DELETE
operation is submitted. When the method is DELETE, the length of the
response content is checked. If the length is 0, an empty list is
returned.

This prevents the method from raising 'requests.exceptions.JSONDecodeError'
due to attempting to parse an empty response as JSON.