Adds validation to add_static_group_members.
A new, optional validate parameter is added to the add_static_group_members method with a default value of False. When set to True, the method will compare the list of members added with the response received from Turbonomic to validate that all members were added successfully. If any were not added, which is what happens if the UUID is invalid, the changes are rolled back and a new error type of VMTGroupValidationError is raised.

Validation is skipped if both the members to add is an empty list and
the group itself is currently empty. While shouldn't happen since there
would be no point, if it did then a KeyError would be raised since the
response will not have a memberUuidList key.