Removes turboauth from the library.
Removes the need for the cryptography library from the library. Updates
the version of requests to 2.28.1. Also bumps the required version of
Python to 3.7 as 3.6 is no longer supported. Removes the entrypoint
statement from the setup file.

Unused imports for math, urlencode, and security (which contained
turboauth) are removed from __init__.